# api/cmd/api/openapi.json
{
  "components": {
    "schemas": {
      "CreateSandboxRequest": {
        "description": "Sandbox create request. CPU/memory may be overridden by the template's baked snapshot dimensions.",
        "properties": {
          "cpu": {
            "description": "Requested vCPUs",
            "type": "integer"
          },
          "disk_gb": {
            "description": "Optional rootfs size in GiB",
            "type": "integer"
          },
          "from_snapshot": {
            "description": "Restore from this snapshot ID instead of booting the template",
            "type": "string"
          },
          "memory_mb": {
            "description": "Requested memory in MiB",
            "type": "integer"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Free-form string metadata; keys starting with secret. are write-only",
            "type": "object"
          },
          "persistent": {
            "description": "true exempts the sandbox from the idle reaper",
            "type": "boolean"
          },
          "template": {
            "description": "Template name (e.g. base, code-interpreter, agent). Public templates or ones owned by the caller's workspace only",
            "type": "string"
          },
          "ttl_seconds": {
            "description": "Idle TTL in seconds; null uses the platform default",
            "type": [
              "integer",
              "null"
            ]
          },
          "volumes": {
            "description": "Durable volumes to attach",
            "items": {
              "$ref": "#/components/schemas/VolumeMount"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "description": "Human-readable error message",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Sandbox": {
        "description": "A Firecracker microVM sandbox. secret.* metadata keys are redacted from all responses.",
        "properties": {
          "boot_mode": {
            "description": "How the sandbox booted",
            "enum": [
              "cold",
              "warm",
              "snapshot",
              null
            ],
            "type": [
              "string",
              "null"
            ]
          },
          "boot_ms": {
            "description": "Boot duration in milliseconds",
            "type": "integer"
          },
          "cpu": {
            "description": "vCPU count (may be overridden to match the template's baked snapshot)",
            "type": "integer"
          },
          "created_at": {
            "description": "Creation timestamp",
            "format": "date-time",
            "type": "string"
          },
          "from_snapshot": {
            "description": "Snapshot ID this sandbox was restored from, if any",
            "type": [
              "string",
              "null"
            ]
          },
          "guest_ip": {
            "description": "Guest IP inside its /30 subnet",
            "type": "string"
          },
          "host_tap": {
            "description": "Host-side TAP/netns identifier",
            "type": "string"
          },
          "id": {
            "description": "Sandbox UUID",
            "type": "string"
          },
          "mac": {
            "description": "Guest NIC MAC address",
            "type": "string"
          },
          "memory_mb": {
            "description": "Memory in MiB (may be overridden to match the template's baked snapshot)",
            "type": "integer"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Free-form string metadata; metadata.workspace is stamped server-side",
            "type": [
              "object",
              "null"
            ]
          },
          "status": {
            "description": "Current lifecycle status",
            "enum": [
              "creating",
              "running",
              "paused",
              "stopping",
              "deleted",
              "failed",
              "hibernated"
            ],
            "type": "string"
          },
          "template": {
            "description": "Template the sandbox was created from (e.g. base, code-interpreter)",
            "type": "string"
          },
          "vsock_cid": {
            "description": "vsock context ID for guest-agent communication",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "SuspendedError": {
        "properties": {
          "detail": {
            "description": "Human-readable explanation",
            "type": "string"
          },
          "error": {
            "description": "Short error code message",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "VolumeMount": {
        "properties": {
          "name": {
            "description": "Volume name",
            "type": "string"
          },
          "read_only": {
            "description": "Mount read-only",
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "pds_... API token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "url": "https://pukucloud.ai"
    },
    "description": "REST API for PukuCloud — Firecracker microVM sandboxes and managed Postgres. Authenticate with a pds_ bearer token from the dashboard's API Tokens page.",
    "title": "PukuCloud API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sandboxes": {
      "get": {
        "operationId": "listSandboxes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Sandbox"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Array of sandboxes owned by the caller's workspace"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "List failed"
          }
        },
        "summary": "List the caller's sandboxes (workspace-scoped; internal pool/tombstone rows filtered out)",
        "tags": [
          "Sandboxes"
        ]
      },
      "post": {
        "operationId": "createSandbox",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sandbox"
                }
              }
            },
            "description": "Sandbox created and running"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed JSON body or invalid template name"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspendedError"
                }
              }
            },
            "description": "Account suspended, or template owned by another workspace"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Create failed"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No available compute node"
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Host memory admission refused (control plane places elsewhere)"
          }
        },
        "summary": "Create a sandbox (snapshot-restore boot, ~180ms p50)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
    "/v1/sandboxes/{id}": {
      "delete": {
        "operationId": "deleteSandbox",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Managed sandbox (a database or hosted app VM) — delete it via its owning feature's API instead"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Delete failed"
          }
        },
        "summary": "Delete a sandbox (irreversible; does not wake a hibernated sandbox first)",
        "tags": [
          "Sandboxes"
        ]
      },
      "get": {
        "operationId": "getSandbox",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sandbox"
                }
              }
            },
            "description": "Sandbox (secret.* metadata keys redacted)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found (or owned by another workspace)"
          }
        },
        "summary": "Get a sandbox (does not wake a hibernated sandbox)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
...
    "/v1/sandboxes/{id}/exec/pty": {
      "get": {
        "operationId": "openPty",
        "parameters": [
          {
            "description": "Sandbox UUID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Initial terminal rows",
            "in": "query",
            "name": "rows",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Initial terminal columns",
            "in": "query",
            "name": "cols",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols — WebSocket PTY session established"
          },
          "404": {
            "description": "Sandbox not found or guest unreachable"
          }
        },
        "summary": "Open an interactive PTY over WebSocket (xterm.js compatible). Client binary frames = stdin; text frames = JSON control ({\"resize\":{\"rows\":N,\"cols\":N}}). Server binary frames = terminal output; text frames = JSON status ({\"exit\":0} or {\"error\":\"...\"}).",
        "tags": [
          "Sandboxes"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://api.pukucloud.ai"
    }
  ]
}